can now have multiple emitters
Emitters now emit particles. and can dynamicly add them my clicking. You can also hold the mouse button down and many many many will be created.
Fixed bug which cause memory issues. In my previous implementations of this easch emitter will be charged with loading in another version of the flame graphic. Now it will be loaded in onece and each particle will be given a refrence to the graphic.